rename queries module to db
