+++ backend/src/db/mod.rs
@@ -1,3 +1,4 @@
+pub struct Db;
 pub mod messages;
 pub mod channels;
 pub mod users;
